requirements.txt
Flask==2.3.3
google-cloud-firestore==2.11.1
python-barcode==0.15.1
Pillow==10.0.0
python-dotenv==1.0.0
